fixing item size and grid count calculation
