+++ app/src/main/java/com/andor/watchit/screens/searchmovie/view/SearchMovieViewMvcImpl.kt
@@ -2,7 +2,6 @@ package com.andor.watchit.screens.searchmovie.view
 
 import android.app.SearchManager
 import android.content.Context
-import android.util.DisplayMetrics
 import android.view.*
 import androidx.appcompat.widget.SearchView
 import androidx.core.widget.ContentLoadingProgressBar
@@ -14,6 +13,7 @@ import com.andor.watchit.R
 import com.andor.watchit.core.inVisible
 import com.andor.watchit.core.visible
 import com.andor.watchit.screens.common.ViewMvcFactory
+import com.andor.watchit.screens.common.helper.Utils
 import com.andor.watchit.screens.common.mvc.BaseObservableViewMvc
 import com.andor.watchit.screens.searchmovie.model.Event
 import com.andor.watchit.screens.searchmovie.view.searchmovieitem.controller.SearchMovieListAdapter
@@ -52,27 +52,13 @@ class SearchMovieViewMvcImpl(
 
             recyclerView.adapter = this.adapter
 
-            val gridCount = getPossibleGridCount()
+            val gridCount = Utils.getPossibleGridCount(context)
 
             recyclerView.layoutManager = GridLayoutManager(context, gridCount)
 
         } else {
             adapter = recyclerView.adapter as SearchMovieListAdapter
         }
-    }
-
-    private fun getPossibleGridCount(): Int {
-        val windowManager = context.getSystemService(Context.WINDOW_SERVICE) as WindowManager
-        val display: Display = windowManager.defaultDisplay
-        val outMetrics = DisplayMetrics()
-        display.getMetrics(outMetrics)
-
-        val density: Float = context.resources.displayMetrics.density
-        val dpWidth = outMetrics.widthPixels / density
-
-        val dimension = context.resources.getDimension(R.dimen.itemWidth)
-        val possibleGridCount = (dpWidth / dimension).toInt()
-        return if (possibleGridCount > 2) possibleGridCount else 2
     }
 
     override fun setSearchBar(
