+++ app/src/main/java/com/andor/watchit/screens/topratedmovielist/view/topratedlistitem/view/TopRatedMovieListItemViewMvcImpl.kt
@@ -22,7 +22,7 @@ class TopRatedMovieListItemViewMvcImpl(
     private var moviePosterImageView: ImageView
 
     init {
-        setRootView(inflater.inflate(R.layout.top_rated_movie_list_item, parent, false))
+        setRootView(inflater.inflate(R.layout.movie_list_item, parent, false))
         moviePosterImageView = findViewById(R.id.moviePosterImageView)
         moviePosterContainer = findViewById(R.id.moviePosterContainer)
     }
